+++ src/RDFRepresentation.ts
@@ -8,10 +8,6 @@ import * as HTTP from "./HTTP";
 export const RDF_CLASS:string = NS.C.Class.RDFRepresentation;
 
 export const SCHEMA:ObjectSchema.Class = {
-	"fileIdentifier": {
-		"@id": NS.C.Predicate.fileIdentifier,
-		"@type": NS.XSD.DataType.string,
-	},
 	"mediaType": {
 		"@id": NS.C.Predicate.mediaType,
 		"@type": NS.XSD.DataType.string,
@@ -23,7 +19,6 @@ export const SCHEMA:ObjectSchema.Class = {
 };
 
 export interface Class extends PersistedDocument.Class {
-	fileIdentifier:string;
 	mediaType:string;
 	size:number;
 
@@ -32,8 +27,7 @@ export interface Class extends PersistedDocument.Class {
 
 export class Factory {
 	static hasClassProperties( object:Object ):boolean {
-		return Utils.hasPropertyDefined( object, "fileIdentifier" )
-			&& Utils.hasPropertyDefined( object, "mediaType" )
+		return Utils.hasPropertyDefined( object, "mediaType" )
 			&& Utils.hasPropertyDefined( object, "size" )
 			&& Utils.hasPropertyDefined( object, "download" );
 	}